DEV: Add tables and models for documentation sidebar indexes, sections, and links (#46)
We want to move docs from live parsing of the index to load from tables instead, this would in the future support better UX around index creation.

This commit creates the tables with appropriate validations. The Index table will replace the category custom field. The migration will come in a subsequent commit as we switch event handlers over at the same time.

This commit does not include the custom field data migration, that would mean the whole plugin has to switch over and make the commit too unmanageable.