FIX: Do not show urls in sidebar but show topic titles instead (#68)
The sidebar URLs sometimes show URLs instead of titles. This is generally hard to reproduce locally, but the fix involves checking if the link text is the same as the href and falling back to the topic title from the database instead.

The IndexStructureRefresher takes in post.cooked, so there's a chance there's some race condition with ProcessPost which updates the topic URL, regardless, the fix above would work.